Settings screen
- Checking if already the social site is registered with other describe
account.

+++ Developer/Components/Classes/View Designs/Modules/Profile/DProfileDetailsViewController.m
@@ -318,7 +318,7 @@
 
 - (void)segmentViewDidSelected:(DSegmentView *)segmentView atIndex:(NSNumber *)index
 {
-    int indexx = [index integerValue];
+    NSInteger indexx = [index integerValue];
     switch (indexx) {
         case 0:
             [self removeViewFromSuperView:_followersList];
